# frontend/openapi/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v2/tags:
    post:
      operationId: TagsController_create
      summary: Créer un nouveau tag.
      description: "

        Ce endpoint permet de créer un tag.


        Information requise :\ 

        - **name** : nom unique du tag

        \    "
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagDto"
      responses:
        "201":
          description: Tag créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        "403":
          description: Accès refusé - Privilège admin requis
      tags: &a7
        - Tags
    get:
      operationId: TagsController_findAll
      summary: Rechercher des tags.
      description: "

        \  Cette route permet de rechercher selon les paramètres fournis :

        \ \ 

        \  - Retourne une liste paginée ou non de tags.

        \  - Filtrage possible par nom.

        \  - Tri possible par :

        \    - `popularité` (défaut)

        \    - `name`

        \    - `createdAt`

        \  - Ordre : `asc` ou `desc`.

        \  "
      parameters:
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
        - name: name
          required: false
          in: query
          description: Recherche sur le nom
          schema:
            example: tag1
            type: string
      responses:
        "200":
          description: Liste des tags trouvés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTagDto"
      tags: *a7
  /api/v2/tags/{id}:
    get:
      operationId: TagsController_findOne
      summary: Rechercher un tag par id.
      parameters:
        - name: id
          required: true
          in: path
          description: ID du tag à trouver
          schema:
            type: string
      responses:
        "200":
          description: Tag trouvé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        "404":
          description: Tag non trouvé
      tags: *a7
    patch:
      operationId: TagsController_update
      summary: Modifier un tag.
      parameters:
        - name: id
          required: true
          in: path
          description: ID du tag à modifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTagDto"
      responses:
        "200":
          description: Tag mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        "403":
          description: Accès refusé - Privilège admin requis
        "404":
          description: Tag non trouvé
      tags: *a7
    delete:
      operationId: TagsController_delete
      summary: Supprimer un tag.
      parameters:
        - name: id
          required: true
          in: path
          description: ID du tag à supprimer
          schema:
            type: string
      responses:
        "204":
          description: Tag supprimé avec succès
        "403":
          description: Accès refusé - Privilège admin requis
        "404":
          description: Tag non trouvé
      tags: *a7
...
  /api/v2/tokens:
    get:
      operationId: TokenController_list
      summary: Récupère tous les tokens de service
      description: Cette méthode permet de récupérer tous les tokens de service.
      parameters: []
      responses:
        "200":
          description: Liste des tokens récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenDto"
      tags: &a10
        - tokens
    post:
      operationId: TokenController_createService
      summary: Crée un nouveau token de service
      description: Cette méthode permet de créer un nouveau token de service.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceTokenDto"
      responses:
        "201":
          description: Token créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExposedTokenDto"
      tags: *a10
  /api/v2/tokens/personal:
    get:
      operationId: TokenController_findPersonal
      summary: Récupère tous les tokens personnels
      description: Cette méthode permet de récupérer tous les tokens personnels.
      parameters: []
      responses:
        "200":
          description: Liste des tokens récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenDto"
      tags: *a10
    post:
      operationId: TokenController_createPersonal
      summary: Crée un nouveau token personnel
      description: Cette méthode permet de créer un nouveau token personnel.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonalTokenDto"
      responses:
        "201":
          description: Token créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExposedTokenDto"
      tags: *a10
  /api/v2/tokens/{id}/regenerate:
    post:
      operationId: TokenController_regenerate
      summary: Régénère un token de service
      description: Cette méthode permet de régénérer un token de service.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateTokenDto"
      responses:
        "200":
          description: Token régénéré avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExposedTokenDto"
      tags: *a10
  /api/v2/tokens/{id}:
    delete:
      operationId: TokenController_delete
      summary: Supprime (révoque) un token de service
      description: Cette méthode permet de supprimer (révoquer) un token de service.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Token supprimé avec succès
      tags: *a10
  /api/v2/tokens/personal/{id}:
    delete:
      operationId: TokenController_deletePersonal
      summary: Supprime (révoque) un token personnel
      description: Cette méthode permet de supprimer (révoquer) un token personnel.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Token supprimé avec succès
      tags: *a10
...
  /api/v2/applications/{applicationId}/technical-debt-info:
    post:
      operationId: ApplicationTechnicalDebtInfoController_create
      summary: Create technical debt info for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnicalDebtInfoDto"
      responses:
        "201":
          description: Technical debt info created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalDebtInfoDto"
        "409":
          description: Technical debt info already exists for this application
      tags: &a21
        - Technical Debt Info
    get:
      operationId: ApplicationTechnicalDebtInfoController_findOne
      summary: Retrieve the technical debt info for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      responses:
        "200":
          description: Technical debt info found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalDebtInfoDto"
      tags: *a21
    patch:
      operationId: ApplicationTechnicalDebtInfoController_update
      summary: Update the technical debt info for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTechnicalDebtInfoDto"
      responses:
        "200":
          description: Technical debt info updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalDebtInfoDto"
      tags: *a21
...
info:
  title: API Référentiel Applications
  description: API pour la gestion des applications
  version: "2.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/auth
          tokenUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          refreshUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          scopes:
            openid: OpenID scope
            profile: Profile scope
      description: OAuth2 authentication using OIDC
    api_key:
      type: apiKey
      in: header
      name: x-refapp-token
      description: Token authentication
  schemas:
...
    TechnicalDebtInfoDto:
      type: object
      properties:
        technicalMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 3.2
          description: Technical maturity score (0-5)
        businessMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
          description: Business maturity score (0-5)
        costMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 2.75
          description: Cost maturity score (0-5)
        id:
          type: string
          example: uuid-example
          description: Unique identifier
        applicationId:
          type: string
          example: uuid-example
          description: Application ID
      required:
        - id
        - applicationId
...
    CreateTagDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 128
          description: Nom unique du tag
          example: securite
      required:
        - name
    TagDto:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique du tag
        name:
          type: string
          minLength: 2
          maxLength: 128
          pattern: /^[a-z._-]+$/
          description: Nom unique du tag
          example: securite
        createdAt:
          type: string
          description: Date de création du tag
          example: 2025-01-01T00:00:00.000Z
      required:
        - id
        - name
        - createdAt
    PaginatedTagDto:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/TagDto"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    UpdateTagDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 128
          description: Nom unique du tag
          example: securite
...
    TokenDto:
      type: object
      properties:
        id:
          type: string
          example: abc123def456ghi789jkl012mno345pq
          description: Identifiant unique du token
        adminLevel:
          enum: &a26
            - 0
            - 10
            - 20
            - 30
          type: number
          example: 10
          description: Niveau d'administration du token
        description:
          type: string
          example: Token pour l'application X
          description: Description du token
        name:
          type: string
          example: Service X
          description: Nom du service ou de l'application utilisant le token
        expiresAt:
          type: string
          example: 2023-12-31T23:59:59.999Z
          description: Date d'expiration du token, au format ISO 8601, maximum 1 an dans
            le futur
        status:
          type: object
          example: active
          description: Statut du token
      required:
        - id
        - description
        - name
        - status
    CreateServiceTokenDto:
      type: object
      properties:
        adminLevel:
          enum:
            - 0
            - 10
            - 20
            - 30
          type: number
          example: 10
          description: Niveau d'administration du token
        description:
          type: string
          example: Token pour l'application X
          description: Description du token
        name:
          type: string
          example: Service X
          description: Nom du service ou de l'application utilisant le token
        expiresAt:
          type: string
          example: 2023-12-31T23:59:59.999Z
          description: Date d'expiration du token, au format ISO 8601, maximum 1 an dans
            le futur
      required:
        - adminLevel
        - description
        - name
    ExposedTokenDto:
      type: object
      properties:
        id:
          type: string
          example: abc123def456ghi789jkl012mno345pq
          description: Identifiant unique du token
        adminLevel:
          enum: *a26
          type: number
          example: 10
          description: Niveau d'administration du token
        description:
          type: string
          example: Token pour l'application X
          description: Description du token
        name:
          type: string
          example: Service X
          description: Nom du service ou de l'application utilisant le token
        expiresAt:
          type: string
          example: 2023-12-31T23:59:59.999Z
          description: Date d'expiration du token, au format ISO 8601, maximum 1 an dans
            le futur
        status:
          type: object
          example: active
          description: Statut du token
        password:
          type: string
          example: xyz987uvw654rst321opq098nml765kj
          description: Valeur du token (ne sera pas stockée, à afficher une seule fois)
      required:
        - id
        - description
        - name
        - status
        - password
    CreatePersonalTokenDto:
      type: object
      properties:
        adminLevel:
          enum:
            - 0
            - 10
            - 20
            - 30
          type: number
          example: 10
          description: Niveau d'administration du token personnel (optionnel). Il ne
            pourra pas être supérieur à celui de l'utilisateur créant le token.
        description:
          type: string
          example: Token pour l'application X
          description: Description du token
        name:
          type: string
          example: Service X
          description: Nom du service ou de l'application utilisant le token
        expiresAt:
          type: string
          example: 2023-12-31T23:59:59.999Z
          description: Date d'expiration du token, au format ISO 8601, maximum 1 an dans
            le futur
      required:
        - description
        - name
    RegenerateTokenDto:
      type: object
      properties:
        expiresAt:
          type: string
          example: 2023-12-31T23:59:59.999Z
          description: Date d'expiration du token, au format ISO 8601, maximum 1 an dans
            le futur
...
    CreateTechnicalDebtInfoDto:
      type: object
      properties:
        technicalMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 3.2
          description: Technical maturity score (0-5)
        businessMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
          description: Business maturity score (0-5)
        costMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 2.75
          description: Cost maturity score (0-5)
    UpdateTechnicalDebtInfoDto:
      type: object
      properties:
        technicalMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 3.2
          description: Technical maturity score (0-5)
        businessMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
          description: Business maturity score (0-5)
        costMaturity:
          type: number
          minimum: 0
          maximum: 5
          example: 2.75
          description: Cost maturity score (0-5)
...
security:
  - oauth2: []
  - api_key: []